+++ changelog.md
@@ -5,6 +5,7 @@ Now requires InsaneLib 1.21.13
 * Enhanced the cooldown mechanic
   * The cooldown of a shield is now based on the time it has been lifted
 * Fixed shield cooldown ignoring Fast Recovery enchantment config option
+* Partially fixed a [vanilla bug](https://bugs.mojang.com/browse/MC/issues/MC-105068) that makes players play the hurt sound when blocking
 
 ## 1.11.9
 * Definitely removed Shield Bash enchantment